Inclusão de instruções no processo

+++ src/GerarPlanilhaGlobal.ts
@@ -11,6 +11,8 @@ import fs = require('fs');
 import fse = require('fs-extra');
 import path = require('path');
 
+import esprima = require('esprima');
+
 var exectimer = require('exectimer');
 const child_process = require('child_process');
 var uuid = require('node-uuid');
@@ -49,6 +51,7 @@ var caminhoOriginal = DiretorioResultados + `/${heuristicas[0]}/original.js`;
 var codigoOriginal = fs.readFileSync(caminhoOriginal, 'UTF8');
 var originalLoc = 0;
 var originalChar = 0;
+var originalIns = 0;
 
 console.log(`${DiretorioBiblioteca}`);
 console.log(`${arquivoRootBiblioteca}`);
@@ -94,6 +97,9 @@ async function Executar() {
     originalLoc = resultadoOriginal.Loc;
     originalChar = resultadoOriginal.Chars;
 
+    var generatedAST = esprima.parse(result.code) as any;
+    originalIns = CountNodes(generatedAST);
+
     resultadosProcessados.push(resultadoOriginal);
 
     //============================================================================================ Rodadas //>
@@ -102,6 +108,7 @@ async function Executar() {
         var heuristica = heuristicas[j];
         var BestLoc = originalLoc;
         var BestChars = originalChar;
+        var BestInst = originalIns;
         var BestResult: TestResults = JSON.parse(JSON.stringify(resultadoOriginal));
 
         BestResult.Heuristic = heuristica;
@@ -127,11 +134,11 @@ async function Executar() {
                 WriteCodeToFile(arquivoRootBiblioteca, CodigoDaRodada);
 
                 var resultadoFinal = await ExecutarTeste(DiretorioBiblioteca, bufferOption, Quantidade);
-                if(!resultadoFinal.passedAllTests){
+                if (!resultadoFinal.passedAllTests) {
                     console.log("Falhou nos testes!")
                     continue;
                 }
-                
+
                 resultadoFinal.Heuristic = heuristica;
 
                 var result = UglifyJS.minify(CodigoDaRodada, uglifyOptions);
@@ -140,8 +147,10 @@ async function Executar() {
 
                 resultadoFinal.Loc = result.code.split(/\r\n|\r|\n/).length;
                 resultadoFinal.Chars = result.code.length;
+                var generatedAST = esprima.parse(result.code) as any;
+                resultadoFinal.Instructions = CountNodes(generatedAST);
 
-                if (resultadoFinal.passedAllTests && resultadoFinal.Chars < BestChars) {
+                if (resultadoFinal.passedAllTests && (resultadoFinal.Chars < BestChars || resultadoFinal.Instructions < BestInst)) {
 
                     var fileContents = fs.readFileSync(caminhoArquivoCVSRodada).toString().replace('sep=,\n', '');
                     if (fileContents.length === 0) {
@@ -160,6 +169,7 @@ async function Executar() {
                     BestResult = resultadoFinal;
                     BestChars = resultadoFinal.Chars;
                     BestLoc = resultadoFinal.Loc;
+                    BestInst = resultadoFinal.Instructions;
                 }
 
                 //Volta a cópia de segurança
@@ -176,7 +186,10 @@ async function Executar() {
     //============================================================================================ Fim //>
     //Volta a cópia de segurança
     fse.copySync(oldLibFilePath, arquivoRootBiblioteca, { "clobber": true });
-    fse.unlinkSync(caminhoArquivoCVSRodadaAlterado);
+
+    if (fse.existsSync(caminhoArquivoCVSRodadaAlterado))
+        fse.unlinkSync(caminhoArquivoCVSRodadaAlterado);
+
     console.log(`Escrever csv com ${resultadosProcessados.length} resultados obtidos`);
 
     EscreverResultadoEmCsv(DiretorioResultados, resultadosProcessados);
@@ -278,7 +291,7 @@ function ShowConsoleResults(result: TestResults) {
     console.log('median:' + result.median);   // median tick duration
 }
 
-/**
+/** 
  * Salva os resultados na raiz
  */
 function EscreverResultadoEmCsv(DiretorioResultados: string, listaResultados: TestResults[]) {
@@ -286,7 +299,7 @@ function EscreverResultadoEmCsv(DiretorioResultados: string, listaResultados: Te
     var newLine: string = '\n';
     //var csvcontent = "sep=;" + newLine;
     var csvcontent = "";
-    csvcontent += "Lib;Heuristic;Trial;Lines;% Improved Loc;Chars;% Improved Chars;Time Spent" + newLine;
+    csvcontent += "Lib;Heuristic;Trial;Lines;% Improved Loc;Chars;% Improved Chars;Instructions;% Improved Instructions;Time Spent" + newLine;
 
     listaResultados.forEach(element => {
         var improvedLoc = ((originalLoc - element.Loc) / originalLoc);
@@ -295,8 +308,10 @@ function EscreverResultadoEmCsv(DiretorioResultados: string, listaResultados: Te
         var improvedChars = ((originalChar - element.Chars) / originalChar);
         improvedChars = improvedChars < 0 ? improvedChars * -1 : improvedChars;
 
+        var improvedInstructions = ((originalIns - element.Instructions) / originalIns);
+        improvedInstructions = improvedInstructions < 0 ? improvedInstructions * -1 : improvedInstructions;
 
-        csvcontent += `${libName};${element.Heuristic};${element.Trial};${element.Loc};${String(improvedLoc).replace('.', ',')};${element.Chars};${String(improvedChars).replace('.', ',')};${element.duration}` + newLine;
+        csvcontent += `${libName};${element.Heuristic};${element.Trial};${element.Loc};${String(improvedLoc).replace('.', ',')};${element.Chars};${String(improvedChars).replace('.', ',')};${element.Instructions};${String(improvedInstructions).replace('.', ',')};${element.duration}` + newLine;
 
     });
 
@@ -379,4 +394,9 @@ function split(line, lineNumber) {
     }
     var parts = line.split(',')
     return [parts[0], parts[1], parts[2], parts[3], parts[4], parts[5], parts[6], parts[7] + ',' + parts[8], parts[9]];
+}
+
+function CountNodes(AST: Object): number {
+    var traverse = require('traverse');
+    return traverse(AST).nodes().length;
 }
